change int to long

+++ week_06/recursive_multistage_reduction/seq.c
@@ -6,24 +6,24 @@ int main(int argc, char **argv) {
 
     srand(time(NULL));
 
-    int n;
-    int count = 0;
+    long n;
+    long count = 0;
 
     printf("N = ");
-    scanf("%d", &n);
+    scanf("%ld", &n);
 
-    int *array = (int *) malloc(sizeof(int) * n);
-    
-    for (int i = 0; i < n; i++) {
+    long *array = (long *) malloc(sizeof(int) * n);
+
+    for (long i = 0; i < n; i++) {
         array[i] = (rand() % 2);
     }
 
-    for (int i = 0; i < n; i++) {
+    for (long i = 0; i < n; i++) {
         if (array[i] == 1)
             count++;
     }
 
-    printf("Count: %d\n", count);
+    printf("Count: %ld\n", count);
     
     return EXIT_SUCCESS;
 }
